Remove unnecessary conditionals for checking API level
